Library

Index

Webviews

Webviews.Webview

Webviews.WindowSizeHint

Base.bind

Base.resize!

Base.run

Base.sizehint!

Webviews.destroy

Webviews.eval!

Webviews.html!

Webviews.init!

Webviews.navigate!

Webviews.terminate

Webviews.title!

Webviews.unbind

Webviews.window_handle

Public Interface

Webviews — Module

Webviews

Julia wrappers for webview, a tiny cross-platform webview library.

Webviews.Webview — Type

Webview(size=(1024, 768); title="", debug=false, size_hint=WEBVIEW_HINT_NONE, unsafe_window_handle=C_NULL)
Webview(width, height; kwargs...)

Create a new webview instance with size and title.

If debug is true, developer tools will be enabled (if the platform supports them).

UNSAFE unsafe_window_handle can be an unsafe pointer to the platforms specific native window handle.

If it's non-null - then child WebView is embedded into the given parent window. Otherwise a new window is created. Depending on the platform, a GtkWindow, NSWindow or HWND pointer can be passed here.

Webviews.WindowSizeHint — Type

WindowSizeHint

Enum to specify the window size hint. Values:

WEBVIEW_HINT_NONE: Width and height are default size.

WEBVIEW_HINT_MIN: Width and height are minimum bounds.

WEBVIEW_HINT_MAX: Width and height are maximum bounds.

WEBVIEW_HINT_FIXED: Window size can not be changed by a user.

Base.run — Method

run(w::Webview)

Run the webview event loop. Runs the main event loop until it's terminated. After this function exits, the webview is automatically destroyed.

Webviews.terminate — Function

terminate(w::Webview)

Stops the main loop. It is safe to call this function from another other background thread.

Note: This function is not working on macOS.

Webviews.destroy — Function

destroy(w::Webview)

Destroys the webview and closes the window along with freeing all internal resources.

Webviews.window_handle — Function

window_handle(w::Webview)

UNSAFE An unsafe pointer to the webviews platform specific native window handle. When using GTK backend the pointer is GtkWindow pointer, when using Cocoa backend the pointer is NSWindow pointer, when using Win32 backend the pointer is HWND pointer.

Webviews.title! — Function

title!(w::Webview, title::AbstractString)

Set the native window title.

Base.resize! — Method

resize!(w::Webview, [size::Tuple{Integer, Integer}]; [hint::WindowSizeHint])

Sets the native window size.

Base.sizehint! — Method

sizehint!(w::Webview, hint::WindowSizeHint)

Sets the native window size hint.

Webviews.navigate! — Function

navigate!(w::Webview, url::AbstractString)

Navigates webview to the given URL. URL may be a data URI, i.e. "data:text/html,<html>...</html>".

Webviews.html! — Function

html!(w::Webview, html::AbstractString)

Set webview HTML directly.

Webviews.init! — Function

init(w::Webview, js::AbstractString)

Injects JavaScript code at the initialization of the new page. Every time the webview will open a the new page - this initialization code will be executed. It is guaranteed that code is executed before window.onload.

Webviews.eval! — Function

eval!(w::Webview, js::AbstractString)

Evaluates arbitrary JavaScript code. Evaluation happens asynchronously, also the result of the expression is ignored. Use bind if you want to receive notifications about the results of the evaluation.

Base.bind — Method

bind(f::Function, w::Webview, name::AbstractString)

Binds a callback so that it will appear in the webview with the given name as a global async JavaScript function. Callback arguments are automatically converted from json to as closely as possible match the arguments in the webview context and the callback automatically converts and returns the return value to the webview.

Note: The callback function must handle a Tuple as its argument.

Webviews.unbind — Function

unbind(w::Webview, name::AbstractString)

Unbinds a previously bound function freeing its resource and removing it from the webview JavaScript context.